+++ src/index.jsx
@@ -41,7 +41,7 @@ import './style.scss';
 import React from 'react';
 
 import appReducer from './reducers';
-import Manager from './Manager/Manager';
+import Environments from './Environments/Environments';
 import Settings from './Settings/Settings';
 import FirstInstallInstructions from './FirstInstall/FirstInstallInstructions';
 
@@ -55,7 +55,7 @@ export default {
     decorateMainView: MainView => ({ viewId }) => (
         <MainView cssClass="main-view">
             <div className="main-view-scroll">
-                <Manager className={showIf(viewId === 0)} />
+                <Environments className={showIf(viewId === 0)} />
                 <Settings className={showIf(viewId === 1)} />
                 <FirstInstallInstructions className={showIf(viewId === 2)} />
             </div>
